refactor(dashboard): split RecordingList into focused components
Extract three siblings out of the 633-line RecordingList:

- RecordingListToolbar (search + sort/density dropdowns + count)
- RecordingRow (per-recording row, in-flight pill, row-actions menu,
  delete confirm)
- PendingUploadRow (optimistic placeholder)

Parent keeps all state, filtering, grouping, infinite-scroll
IntersectionObserver, and the useImperativeHandle keyboard nav. Row
refs are now registered via a parent-supplied registerRef callback so
next()/prev() scroll-into-view keeps working.

Public surface of RecordingList is unchanged: same props, same exported
types (SortOrder, ListDensity, PendingUpload, RecordingListHandle),
same behavior. SortOrder/ListDensity now live in the toolbar file and
PendingUpload in pending-upload-row.tsx; both are re-exported from
recording-list.tsx for existing callers.
